% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallistoRes.R
\name{addKallistoResToTable}
\alias{addKallistoResToTable}
\title{Adds the output kallisto directory to the original datafile.}
\usage{
addKallistoResToTable(dataFile, countDat)
}
\arguments{
\item{dataFile}{MORE}

\item{countDat}{MORE}
}
\value{
dataframe MORE
}
\description{
GOES HERE
}
\seealso{
\code{\link{collectKallistoCounts}}
}

