% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotation based on GO terms}
\usage{
annotate(filter = c(Fold = 2, FDR = 10), id_type = "gene", CLSZ = 2,
  cutoff = 0.01)
}
\arguments{
\item{filter}{Named vector with filter cutoffs of format c(Fold=2, FDR=10) where Fold refers to the fold change cutoff (unlogged) and FDR to the p-value cutoff, default value is Fold=2, FDR=10.}

\item{id_type}{specifies type of IDs in input, can be assigned gene or affy, default is gene.}

\item{CLSZ}{minimum gene set (cluster) size to consider. Gene sets below this cutoff will be ignored, default value is 2.}

\item{cutoff}{p-value cutoff for GO terms to show in result data.frame, default is 0.01.}
}
\description{
Provides user with GO annotation and graphs for all the three GO terms i.e., Molecular Function, Biological Process and Cellular Components
}
\examples{

\dontrun{
annotate()
annotate(filter=c(Fold=2, FDR=70), id_type="gene", CLSZ=2, cutoff=0.01)
}

}
\seealso{
\code{\link{useMart}}
}
