% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\title{Function "annotation"}
\usage{
annotation(genes, filter)
}
\arguments{
\item{genes}{a vector of gene IDs such as ensembl ID, gene name}

\item{filter}{a string that is used as the base for extraction of gene information}
}
\value{
a data frame with corresponding attributes of the input gene IDs
}
\description{
This function creates gene annotation file using provided file and filter
}
