% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationII.R
\name{annotationII}
\alias{annotationII}
\title{Function "annotationII"}
\usage{
annotationII(genes, dataSet1 = "hsapiens_gene_ensembl",
  dataSet2 = "mmusculus_gene_ensembl", filter1 = c("mgi_symbol"),
  filter2 = c("external_gene_name", "ensembl_gene_id", "entrezgene_id"))
}
\arguments{
\item{genes}{a vector of gene IDs such as ensembl ID, gene name}

\item{filter}{a string that is used as the base for extraction of gene information}
}
\value{
a data frame with corresponding attributes of the input gene IDs
}
\description{
This function transfers genes from one species to another
}
