% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{availableReferences}
\alias{availableReferences}
\title{A number of transcriptome references have already been built, or use with \code{kallistoQuant}}
\usage{
availableReferences()
}
\arguments{
\item{None}{No parameters}
}
\value{
A list of available transcriptomes (.idx files)
}
\description{
DESCRIPTION GOES HERE
}
\examples{
availableReferences()
 
}

