% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataFile.R
\name{checkDataFile}
\alias{checkDataFile}
\title{Checks that the format of the "data file" (which is used as input in many functions) is correct.}
\usage{
checkDataFile(x)
}
\arguments{
\item{x.}{A data frame containing: the name of the file to be processed, whether it is single or paried-end data, the sample and replicate ID, and (optional) the name of an output file which results from the \code{runQAandFilter} function will be written to.}
}
\value{
A data frame which has been modified - the column names and classes have been verified (and corrected if necessary), all white space has been removed, observations relating to files which do not exist in the directory have been removed. The function also prints out any duplicated files.
}
\description{
The function takes as input a data file (see \code{datafileTemplate}), and checks whether it is in a format acceptable to be used as input to other functions.
}
\seealso{
\code{datafileTemplate}
}

