% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallistoRes.R
\name{collectKallistoCounts}
\alias{collectKallistoCounts}
\title{Collects raw counts of gene (or transcript) data generated by kallisto.}
\usage{
collectKallistoCounts(baseDir = "./kallistoResults")
}
\arguments{
\item{baseDir}{A character string of the directory where kallisto results are stored. For example, baseDir = "kallistoResults" will search "kallistoResults/*/abundance.tsv".}
}
\value{
matrix. A matrix of (rounded) raw counts of genes (or transcripts) for each sample.
}
\description{
Extracts raw counts from each of the "abundance.tsv" files generated by kallisto (\url{https://pachterlab.github.io/kallisto/about}) expression analysis. The counts are put into a matrix (genes by sample), and rounded. This step collects the raw count data ready for further expression analysis, or for differential expression, using \code{\link{DESeq2}}.
}
\seealso{
kallisto (\url{https://pachterlab.github.io/kallisto/about})
}

