% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallistoRes.R
\name{collectKallistoCountsTXItoGene}
\alias{collectKallistoCountsTXItoGene}
\title{Collects raw counts of genes, based on transcript data generated by kallisto.}
\usage{
collectKallistoCountsTXItoGene(baseDir = "./kallistoResults")
}
\arguments{
\item{baseDir}{A character string of the directory where kallisto results are stored. For example, baseDir = "kallistoResults" will search "kallistoResults/*/abundance.tsv".}
}
\value{
matrix. A matrix of raw counts of genes for each sample.
}
\description{
Extracts raw counts of transcripts from each of the "abundance.tsv" files generated by kallisto (\url{https://pachterlab.github.io/kallisto/about}) expression analysis and converts this to gene counts. The counts are put into a matrix (genes by sample). This step collects the raw count data ready for further expression analysis, or for differential expression, using \code{\link{DESeq2}}.
}

