% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaComp.R
\name{compareQA}
\alias{compareQA}
\title{Compares quality before and after filtering fastq files.
Porvides a graphic of the change in quality of fastq files, before and after running the \code{\link{filterBadSeqs}} function.}
\usage{
compareQA(dataFile, prefiltData, postfiltData)
}
\arguments{
\item{dataFile}{An R data frame with the data to be processed. The R object is a standard format, and must contain the following headings: File, PE, Sample, Replicate, FilteredFile. More information about the file is available at \code{datafileTemplate}. The \code{\link{checkDataFile}} function should be run on the file before use.}

\item{prefiltData}{A FastqQA object resulting from \code{\link{runQA}}. The object contains QA data on fastq files before the filtering procedure (\code{\link{filterBadSeqs}}).}

\item{postfiltData}{A FastqQA object resulting from \code{\link{runQA}}. The object contains QA data on fastq files after the filtering procedure (\code{\link{filterBadSeqs}}).}
}
\value{
a datafrome containing densities and quality scores for each fastq file (both before and after filtering). It also outputs a graph which shows comparisons of quality before and after filtering.
}
\description{
Compares quality before and after filtering fastq files.
Porvides a graphic of the change in quality of fastq files, before and after running the \code{\link{filterBadSeqs}} function.
}

