% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{dataPreprocess}
\alias{dataPreprocess}
\title{Function "dataPreprocess"}
\usage{
dataPreprocess(dataInfo, dataMatrix, pcg_ense, condition, attr1, attr2,
  type1, type2, Order)
}
\arguments{
\item{dataInfo}{a data frame which contains clinical information of data}

\item{dataMatrix}{a matrix consisting of gene expression counts}

\item{pcg_ense}{a character vector of protein-coding ensembl gene IDs}

\item{condition}{a character vector composed of all different group conditions}

\item{attr1}{a string character indicating the column name for "Sample ID" in the clinical data}

\item{attr2}{a string character indicating the desired group name in the clinical data}

\item{type1}{a string character representing one group condition}

\item{type2}{a string character representing the other group condition}

\item{Order}{a logical variable indicating whether or not to use decreasing order to resort the data information}
}
\value{
a list consisting of numbers of each group condition, processed clinical data information and ready-to-use gene expression matrix.
}
\description{
This function preprocesses clinical data and gene expression matrix.
}
