% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{deseq2}
\alias{deseq2}
\title{Function "deseq2"}
\usage{
deseq2(condition, n1, n2, N, countsmatrix, M)
}
\arguments{
\item{condition}{a character vector composed of all different group conditions}

\item{n1}{a numeric variable indicating the number of samples for one of the group condition}

\item{n2}{a numeric variable indicating the number of samples for the other group condition}

\item{N}{a numeric variable indicating the threshold for filtering low counts genes}

\item{countsmatrix}{gene expression matrix with raw counts}

\item{M}{a numeric variable indicating the number of cores for parallel computing}
}
\value{
a list consisting of the DESeq data object and a data frame with all differential gene expression analysis results.
}
\description{
This function Perform Differential Gene Expression Analysis using DESeq2.
}
