% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2Res.R
\name{deseq2Res}
\alias{deseq2Res}
\title{Function "deseq2Res"}
\usage{
deseq2Res(countsmatrix, res, geneSymbol, index, group, dds, Date)
}
\arguments{
\item{countsmatrix}{raw counts matrix}

\item{res}{the result data frame from DESeq2}

\item{geneSymbol}{a data frame with all gene annotation information}

\item{index}{a character used to merge the result and gene annotation file}

\item{group}{a string character indicating the group condition}

\item{dds}{a DESeq data object retrieved from DESeq2 used for accessing normalized counts matrix}

\item{Date}{a Date object obtained from Sys.Date}
}
\value{
a list consisting of the sorted DESeq result, differentially expressed gene lists, raw and normalized counts matrices.
}
\description{
This function saves the differential gene expression analysis results and retrieves raw and normalized counts matrices.
}
