% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblTOname_I.R
\name{ensemblTOname_I}
\alias{ensemblTOname_I}
\title{Function "ensemblTOname_I"}
\usage{
ensemblTOname_I(res_anno, raw, N, condition, n1, n2)
}
\arguments{
\item{res_anno}{a data frame with the result of DESeq along with gene annotation}

\item{raw}{the raw counts matrix}

\item{N}{a numeric value representing the number of samples}

\item{condition}{a string character representing the group condition}

\item{n1}{a numeric variable indicating the number of samples for one of the group condition}

\item{n2}{a numeric variable indicating the number of samples for the other group condition}
}
\value{
a ExpressionSet object of counts matrix based on gene names with group condition information
}
\description{
This function should be called after DESeq, it transfers counts matrix based on ensembl IDs to matrix
based on gene names and takes average for duplicated gene names
}
