% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezTOname.R
\name{entrezTOname}
\alias{entrezTOname}
\title{Function "entrezTOname"}
\usage{
entrezTOname(geneSymbol, geneSet)
}
\arguments{
\item{geneSymbol}{a data frame with all gene annotation information}

\item{geneSet}{a list of pathways with entrezgene IDs, gene set}
}
\value{
a list of pathways with gene names
}
\description{
This function transfers the entrez gene IDs in geneSet to gene names
make sure the input geneSet is a list of pathways with character type entrez gene IDs
}
