% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteR.R
\name{filteR}
\alias{filteR}
\title{Trimming and filtering}
\usage{
filteR(a = "4", b = 36)
}
\arguments{
\item{a}{a character(1) with nchar(a) == 1L giving the letter at or below which a nucleotide is marked as failing. The default is "4" (phred score 20).}

\item{b}{drop reads that are less than the number of nucleotides given in b. Default is 36nt}
}
\description{
trim and filter raw reads containing N, trim as soon as 2 of 5 nucleotides has quality encoding less than "4" (phred score 20) and drop reads that are less than 36nt
}
\details{
Filters and trims each user selected raw data file (in .fastq.gz format), based on three parameters, raw reads containing N are removed, reagins having a phred score less than the value defined by user or less tha a phred score of 20 (by default) is trimmed out. As well as sequence reads that have no of bases less than the value given by user or less than 36nt (default value) is removed.
}
\examples{

\dontrun{
filteR()
filteR(a = "4", b = 38)
}

}
\seealso{
\code{\link{filterFastq}}
}
