% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqcR.R
\name{fqcR}
\alias{fqcR}
\title{quality check report}
\usage{
fqcR(samples = TRUE, no = 1e+06, groups = NULL, tops = 10,
  pairs = NULL, worker = 1)
}
\arguments{
\item{samples}{It reads a random sample from files if this parameter is TRUE.}

\item{no}{Number of sequences to read from each input file. This represents sample size if 'sample' parameter is TRUE, if not represents the chunk size to read on each iteration. Default is read a sample of one million sequences from each input file.}

\item{groups}{group name for each input file.}

\item{tops}{number of top over-represented reads. Default is 10 reads.}

\item{pairs}{combination of files for paired-end reads. By default, all input files are treated as single-end. For paired-end, please define a vector of numbers where two index with the same value represent a pair. Examples, single-end c(1,2,3,4) and paired-end c(1,1,2,2).}

\item{worker}{number of parallel workers.Default is 1.}
}
\description{
takes fastq.gz file present within the data folder as input and provides the user with various graph related to the quality of the reads
}
\details{
Multiple input files present in the folder data of the workspace are processed in parallel using the function of Rqc. The result generated is saved in folder results/QC-report. All the images generated are stored in jpeg format and also a html report is also generated that can be opened in browser for analysis.
}
\examples{

\dontrun{
fqcR()
fqcR(pair = c(1,1,2,2,3,3,4,4,5,5,6,6))
fqcR(pair = c(1,1,2,2,3,3,4,4),  worker = multicoreWorkers())
fqcR(pair = c(1,2,3,4))
}

}
\seealso{
\code{\link{rqc}}
}
