% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frqc.R
\name{frqc}
\alias{frqc}
\title{qc}
\usage{
frqc()
}
\description{
reports basic quality and summary statistics on FASTQ files, including base and quality distribution by position, sequence length distribution, k-mers by position
}
\examples{

\dontrun{
frqc()
}

}
\seealso{
\code{\link{filterFastq}}
}
