% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp.R
\name{gene_exp}
\alias{gene_exp}
\title{Differential Gene Expression plots}
\usage{
gene_exp(sysma, mytargets, type = "SYSargs", x, y, con)
}
\arguments{
\item{sysma}{path to 'param' file; file structure follows a simple name/value syntax that converted into JSON format; for details about the file structure see sample files provided by package. Assign NULL to run the pipeline without 'param' file. This can be useful for running partial workflows, e.g. with pregenerated BAM files.}

\item{mytargets}{path to targets file}

\item{type}{type="SYSargs" returns SYSargs, type="json" returns param file content in JSON format (requires rjson library)}

\item{x}{used for ploting the data to specify weather the data is paired-end or single end}

\item{y}{used for plotting the data to indicate the count of data}

\item{con}{a formula which expresses how the counts for each gene depend on the variables in colData}
}
\description{
write a desc
}
\examples{

\dontrun{
pairs <- c( "paired-end", "paired-end", "paired-end","paired-end","paired-end","paired-end")
count <- c(1,1,2,2,3,3)
condition <- factor(c("N1","N1","N2","N2","M1","M1","M2","M2","M3","M3","M4","M4","U1","U1"))
gene_exp(sysma="./param/tophat.param", mytargets="targetsPE.txt", x = pairs, y = count, con = condition)
}

}
\seealso{
\code{\link{edgeR}}
}
