% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEXP.R
\name{geneEXP}
\alias{geneEXP}
\title{Function "geneExp"}
\usage{
geneEXP(res_anno, geneEX, condition, N, n1, n2)
}
\arguments{
\item{res_anno}{a data frame with the result of DESeq along with gene annotation}

\item{geneEX}{normalized gene expression matrix based on ensembl IDs}

\item{condition}{a string character representing the group condition}

\item{N}{a numeric variable indicating the number of total samples}

\item{n1}{a numeric variable indicating the number of samples for one of the group condition}

\item{n2}{a numeric variable indicating the number of samples for the other group condition}
}
\value{
a gene expression object with matrix based on entrezgene IDs
}
\description{
This function constructs expression set for GSEA and ssGSEA
}
