% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSelect.R
\name{geneSelect}
\alias{geneSelect}
\title{Function "geneSelect"}
\usage{
geneSelect(geneES, type, ratio)
}
\arguments{
\item{geneES}{an ExpressionSet object of gene expression matrix with rows of genes and columns of samples}

\item{type}{a character variable from one of "CV"(coefficient of variation), "IQR"(interquantile range), "MAD"(madian absolute deviation).}

\item{ratio}{a numeric variable indicating how much ratio of genes will be selected.}
}
\value{
an ExpressionSet object of selected gene expression matrix.
}
\description{
This function selects most valuable genes based on different methods
}
