% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionalityReduction.R
\name{GetPCA}
\alias{GetPCA}
\title{GetPCA}
\usage{
GetPCA(object = NULL, dims = 10, verbose = FALSE, ...)
}
\arguments{
\item{object}{Seurat object.}

\item{dims}{Dimensions to obtain.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
PCA representation.
}
\description{
GetPCA
}
