% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionalityReduction.R
\name{GetUMAP}
\alias{GetUMAP}
\title{GetUMAP}
\usage{
GetUMAP(
  object = NULL,
  dims = 10,
  reduction = "pca",
  PCA = TRUE,
  scale = TRUE,
  seed = 42,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{dims}{Dimensions to use.}

\item{reduction}{Reduction to use.}

\item{PCA}{Obtain PCA.}

\item{scale}{Scale the data.}

\item{seed}{Set a random seed. By default, sets the seed to 42.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
UMAP representation.
}
\description{
GetUMAP
}
