% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gseaI}
\alias{gseaI}
\title{Function "gseaI"}
\usage{
gseaI(res_anno, geneEX, geneSet, condition, N, n1, n2, Min, Max)
}
\arguments{
\item{res_anno}{a data frame with the result of DESeq along with gene annotation}

\item{geneEX}{normalized gene expression matrix based on ensembl IDs}

\item{geneSet}{a gene set list, consisting of entrezgene IDs}

\item{condition}{a string character representing the group condition}

\item{N}{a numeric variable indicating the number of total samples}

\item{n1}{a numeric variable indicating the number of samples for one of the group condition}

\item{n2}{a numeric variable indicating the number of samples for the other group condition}

\item{Min}{a numeric variable representing the minimum size of pathway accepted for GSEA}

\item{Max}{a numeric variable representing the maximum size of pathway accepted for GSEA}
}
\value{
a list of gene rank based on entrezgene ID, result of GSEA and an ExpressionSet object of normalized counts matrix based on entrezgene IDs.
}
\description{
This function performs GSEA analysis.
}
