% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaResult.R
\name{gseaResult}
\alias{gseaResult}
\title{Function "gseaResult"}
\usage{
gseaResult(geneSymbol, gseaRes, group, gs, Date)
}
\arguments{
\item{geneSymbol}{a data frame with all gene annotation information}

\item{gseaRes}{the result data frame from GSEA}

\item{group}{a string character indicating the group condition}

\item{gs}{a string character indicating the name of gene set}

\item{Date}{Date object obtained from Sys.Date}
}
\value{
a list of sorted GSEA result and significant pathways with different FDR controls.
}
\description{
This function stores customized results from GSEA
}
