% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{import_dataset}
\alias{import_dataset}
\title{loading data}
\usage{
import_dataset()
}
\description{
copies the data needed to run the example form the system folder to the current working directory
}
\examples{

\dontrun{
import_dataset()
}

}
\seealso{
\code{\link{file.copy}}
}
