% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{kallistoQuant}
\alias{kallistoQuant}
\title{Kallisto quantifies transcript abundance from RNA-Seq data.}
\usage{
kallistoQuant(dataFile, preFilt = FALSE, refIndex,
  refIndexFromPackage = FALSE, bootstrap = 0, fragmentLength = 200,
  fragmentSD = 10, biasCor = FALSE)
}
\arguments{
\item{refIndex}{A character string of the name of the index file (usually, .idx) against which the fastq files are pseudoaligned. This file can be generated from a fasta-formatted transcriptome file using \code{\link{kallistoIndex}}.}

\item{refIndexFromPackage}{A logical, false by default. If the user would like to use an index provided by the package (these can be viewed using \code{availableReferences()}, he/she should specifiy "true" here. Otherwise, the function will assume the index is provided by the user.}

\item{bootstrap}{An integer specifying the number of times to bootstrap. The output will provide a measure of variance.}

\item{fragmentLength}{An integer. Estimated fragment length. Only required for single-end data (for paired-end data, Kallisto is able to calculate the fragment length). Default is 200 bp.}

\item{fragmentSD}{A numeric. The standard deviation of the fragment length. Only required for single-end data (for paired-end data, Kallisto is able to calculate the standard deviation). Default is 10 bp.}

\item{file1}{A character string of the name of the RNA-Seq data file (fastq.gz) to be processed.}

\item{file2}{A character string of the RNA-Seq data file (fastq.gz) to be processed - in the case there is paired-end data.}

\item{pairedEnd}{is a logical. If \code{true} (default), a paired end protocol is chosen (for this, the file1 and file2 parameters must be specified). If \code{false}, a single end protocol will be run, and only file1 will be processed.}
}
\value{
A data frame of the estimated abundances of each transcript specified in the input file(s). The data frame is also saved to a folder, which is given the title of the files, exlcuding the extensions. The folder contains these abundances (in text format and compressed format), as well as information about the run.
}
\description{
The data is broken down into k-mers and each k-mer is pseudoaligned to k-mers in the index.
Because kallisto doesn't rely on full alignment, it is much quicker than other methods, without losing accuracy.
}

