% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{LegendPosition}
\alias{LegendPosition}
\title{LegendPosition}
\usage{
LegendPosition(position = "right")
}
\arguments{
\item{position}{Legend position}
}
\value{
Theme with the legened in the selected position. Available positions are :"none", "right", "left", "top", and "bottom".
}
\description{
LegendPosition
}
