% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{myFeaturePlot}
\alias{myFeaturePlot}
\title{myFeaturePlot}
\usage{
myFeaturePlot(
  df = NULL,
  feature = NULL,
  text_size = 20,
  point_size = 2,
  alpha = 1,
  low_color = "gray",
  high_color = "tomato",
  legend_position = "right",
  legend_point_size = NULL,
  order = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame containing the PCA dimensions to plot and the feature to use as colors. PCA data columns should be named as "Dim1" and "Dim2"}

\item{feature}{Feature to use as color scale.}

\item{text_size}{Plot text size}

\item{point_size}{Plot point size}

\item{alpha}{Point transparency}

\item{low_color}{Scale low color}

\item{high_color}{Scale high color}

\item{legend_position}{The position of the legend. It can take values as "top", bottom", "left", "right", or "none.}

\item{legend_point_size}{Legend point size.}

\item{order}{Whether the points should be ordered.}

\item{...}{Arguments passed to other methods.}
}
\value{
A ggplot object containing the feature plot.
}
\description{
Function to plot a feature
}
