% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{myPCAplot}
\alias{myPCAplot}
\title{myPCAplot}
\usage{
myPCAplot(
  df = NULL,
  group_by = NULL,
  text_size = 20,
  point_size = 2,
  alpha = 1,
  legend_position = "right",
  legend_point_size = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame containing the PCA dimensions to plot and the labels to use as colors. PCA data columns should be named as "Dim1" and "Dim2"}

\item{group_by}{Labels to use as colors.}

\item{text_size}{Plot text size}

\item{point_size}{Plot point size}

\item{alpha}{Point transparency}

\item{legend_position}{The position of the legend. It can take values as "top", bottom", "left", "right", or "none.}

\item{legend_point_size}{Legend point size.}

\item{...}{Arguments passed to other methods.}
}
\value{
A ggplot object containing the PCA plot.
}
\description{
Function to plot PCA
}
