% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameTOentrez.R
\name{nameTOentrez}
\alias{nameTOentrez}
\title{Function "nameTOentrez"}
\usage{
nameTOentrez(geneSymbol, geneSet)
}
\arguments{
\item{geneSymbol}{a data frame with all gene annotation information}

\item{geneSet}{a list of gene set based on gene names}
}
\value{
a list type gene set based on entrezgene IDs.
}
\description{
This function transfers the name of genes in geneSet to entrez gene IDs,
make sure the input geneSet is a list of pathways with "character" type of gene names
}
