% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESplot.R
\name{NESplot}
\alias{NESplot}
\title{Function "NESplot"}
\usage{
NESplot(gseaRes, pval, title, type, H, W, Date)
}
\arguments{
\item{gseaRes}{the result from GSEA}

\item{title}{a string character indicating the name of geneset}

\item{type}{a character indicating the group condition}

\item{H}{height for ES plot}

\item{W}{width for ES plot}

\item{Date}{an Date object}
}
\value{
a ggplot object of the nes plot
}
\description{
This function generates an normalized enrichment score plot.
}
