% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Function "normalization"}
\usage{
normalization(res, geneLGC)
}
\arguments{
\item{res}{unnormalized gene expresison matrix based on ensembl IDs}

\item{geneLGC}{a data frame with gene length and GC content information}
}
\value{
a list consisting of normalized gene expression matrix.
}
\description{
This function performs gene expression normalization.
}
