% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Function "plotDensity"}
\usage{
plotDensity(info, ssgseaRes, axis1 = "UPssgsea", axis2 = "DOWNssgsea",
  type, Date)
}
\arguments{
\item{info}{a data frame with required clinical information}

\item{ssgseaRes}{the result matrix of ssGSEA}

\item{axis1}{a character variable indicating the label for x axis}

\item{axis2}{a character variable indicating the label for y axis}

\item{type}{type of data, group information}

\item{Date}{a Date object obtained from Sys.Date}
}
\value{
the density plot object
}
\description{
This function generates scatter plot with marginal density distributions
}
