% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotElbowPCA}
\alias{PlotElbowPCA}
\title{PlotElbowPCA}
\usage{
PlotElbowPCA(
  object = NULL,
  dims = 30,
  pSize = 4,
  lSize = 0.5,
  alpha = 0.6,
  selDim = NULL,
  text_size = 20
)
}
\arguments{
\item{object}{Seurat object with calculated PCA.}

\item{dims}{Number of dimensions to plot.}

\item{pSize}{Point size.}

\item{lSize}{Line size.}

\item{alpha}{Point transparency.}

\item{selDim}{Selected dimension.}

\item{text_size}{Plot's text size.}
}
\value{
A ggplot object containing the elbow plot.
}
\description{
PlotElbowPCA
}
