% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Function "plotGSEA"}
\usage{
plotGSEA(data_info, gseaRes, gs, gs_name, geneRank, gene.esI, title, K,
  type, attr1, attr2, conditionI, conditionII, N, H1, W1, H2, W2, H3, W3)
}
\arguments{
\item{data_info}{data frame containing clinical information}

\item{gseaRes}{the result from GSEA}

\item{gs}{geneset "CORALINE Signature" generated from our cohort}

\item{gs_name}{the geneset used for GSEA based on gene names}

\item{geneRank}{the gene ranking used for GSEA}

\item{gene.esI}{expression set of normalized gene counts matrix based on gene names}

\item{title}{a string character indicating the name of geneset}

\item{K}{number of top enriched/significant pathways to be plotted}

\item{type}{a character indicating the group condition}

\item{attr1}{a character indicating the column name of sample IDs in the clinical data}

\item{attr2}{a character indicating the column name of desired group in the clinical data}

\item{conditionI}{a character vector indicating the primary group condition}

\item{conditionII}{a character vector indicating the secondary group condition}

\item{N}{a variable indicating one of the value from attr2}

\item{H1}{height for plotting significant pathways}

\item{W1}{width for plotting significant pathway}

\item{H2}{height for plotting enriched pathways}

\item{W2}{width for plotting enriched pathways}

\item{H3}{height for ES plot}

\item{W3}{width for ES plot}
}
\value{
a list of plot object
}
\description{
This function generates variety of plots from GSEA result
}
