% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap_POI.R
\name{plotHeatmap_POI}
\alias{plotHeatmap_POI}
\title{Function "plotHeatmap_POI"}
\usage{
plotHeatmap_POI(condition = c("mutant", "wildtype"), geneES, group,
  palette1, palette2 = c("#A6CEE3", "#1F78B4", "#B2DF8A", "#33A02C",
  "#FB9A99", "#E31A1C", "#FDBF6F", "#FF7F00", "#CAB2D6", "#6A3D9A"),
  groupName = "Pathway", legendName = "Gene expression", POI, GOI,
  rowName = TRUE, distance = "spearman", method = "ward.D2", title,
  s1 = 15, s2 = 4, s3 = 11, s4 = 6, s5 = 12, H = 11, W = 28,
  Date)
}
\arguments{
\item{condition}{a vector of characters representing all groups of samples, default is c("mutant", "wildtype").}

\item{geneES}{an ExpressionSet object representing the gene expression matrix, at least two columns in pData(geneES) is required,
"SampledID" indicating samples and "Group" indicating which condition the samples are in.}

\item{group}{a data frame with each column a different column annotation which needs to be added to the heatmap, rownames correspond to sample ID.}

\item{palette1}{a named list with color palettes for each annotation, each element in the list should be a named vector.}

\item{palette2}{a character type vector representing the colors used for distinguishing "POI", default is
c("#A6CEE3", "#1F78B4", "#B2DF8A", "#33A02C", "#FB9A99", "#E31A1C", "#FDBF6F", "#FF7F00", "#CAB2D6", "#6A3D9A").}

\item{groupName}{a character variable indicating the POI name, default is "Pathway".}

\item{legendName}{a character variable indicating the legend name, default is "Gene expression".}

\item{POI}{a character type vector consisting of names of selected pathways.}

\item{GOI}{a list consisting of gene names in the pathways of "POI"(can be all or partial genes).}

\item{rowName}{a boolean variable indicating whether to show the row names. Default is TRUE.}

\item{distance}{a character variable indicating which distance measure to be used in clustering, either one of
"euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall",
or a dist object. Default is "spearman".}

\item{method}{a character variable indicating the clustering method used, pass to hclust. Default is "ward.D2".}

\item{title}{a character variable containing information for filename.}

\item{s1}{a numerical variable indicating the fontsize of legend title. Default is 15.}

\item{s2}{a numerical variable indicating the fontsize of row names. Default is 4.}

\item{s3}{a numerical variable indicating the fontsize of column names. Default is 11.}

\item{s4}{a numerical variable indicating the height and width of grid in defined unit of column legend. Default is 6.}

\item{s5}{a numerical variable indicating the width of the dendrogram for columns and rows. Default is 12.}

\item{H}{a numerical variable indicating the height of saved figure. Default is 11.}

\item{W}{a numerical variable indicating the width of save figure. Default is 28.}

\item{Date}{a Date object obtained from Sys.Date}
}
\value{
a "HeatmapList" class object.
}
\description{
This function generates heatmap for the genes from selected pathways(pathway of interest)
}
