% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayEnrichment.R
\name{PlotKEGG}
\alias{PlotKEGG}
\title{PlotKEGG}
\usage{
PlotKEGG(
  df = NULL,
  type = "double",
  ntop = 20,
  color_by = "Direction",
  point_size = 3,
  text_size = 15
)
}
\arguments{
\item{df}{Data frame containing KEGG pathways}

\item{type}{Type of selected p-values. Available options are: "double", negatives and positive, "positive", and "negative".}

\item{ntop}{Number of top values.}

\item{color_by}{Label to color points.}

\item{point_size}{Size of points}

\item{text_size}{Size of text in the plot.}
}
\value{
A ggplot object.
}
\description{
PlotKEGG
}
