% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNES.R
\name{plotNES}
\alias{plotNES}
\title{Function "plotNES"}
\usage{
plotNES(gseaRes, N, gs, fdr, s, v, h, W, H, Date)
}
\arguments{
\item{gseaRes}{a list of gsea results for different cohorts}

\item{N}{a numeric variable indicating the number of cohorts being used}

\item{gs}{a character variable indicating the name of genesets used}

\item{fdr}{a numeric variable indicating the FDR cutoff value}

\item{s}{a numeric variable representing the size of labels used in the plot}

\item{v}{a numeric variable representing the value of vjust, vertical position of the labels}

\item{h}{a numeric variable representing the value of hjust, horizontal position of the labels}

\item{W}{a numeric variable representing the width of the figure}

\item{H}{a numeric variable representing the height of the figure}

\item{Date}{a character variable returned from Sys.Date()}
}
\value{
a list of ggplot object of the comparison plot of direction-consistent pathways, up and down-regulated pathways
}
\description{
This function plots a comparison pathway barplot for selected cohorts and chosen gene sets. Only pathways that
}
