% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcanoII.R
\name{plotVolcanoII}
\alias{plotVolcanoII}
\title{Function "plotVolcanoII"}
\usage{
plotVolcanoII(fc, pval, res, attr1, attr2, mytitle, group, UPgenes,
  DWgenes, xmin, xmax, ymin, ymax, W, Date)
}
\arguments{
\item{fc}{a numeric variable representing the log2FoldChange cutoff}

\item{pval}{a numeric variable representing the adjusted p value cutoff}

\item{res}{a data frame containing the differential gene expression analysis result(log2FoldChange and padj)}

\item{attr1}{a character indicating column name of log2FoldChange}

\item{attr2}{a character indicating column name of adjusted p value}

\item{mytitle}{a string character indicating the title of the plot}

\item{group}{a string character indicating the group condition}

\item{UPgenes}{"up-regulated" pathway in the given geneset}

\item{DWgenes}{"down-regulated" pathway in the given geneset}

\item{xmin}{the minimum of x axis}

\item{xmax}{the maximum of x axis}

\item{ymin}{the minimum of y axis}

\item{ymax}{the maximum of y axis}

\item{W}{a numeric variable indicating the width of the plot}

\item{Date}{a Date object obtained from Sys.Date}
}
\value{
a list consisting of normalized gene expression matrix.
}
\description{
This function generates volcano plot
}
