% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCorrection.R
\name{RunHarmony}
\alias{RunHarmony}
\title{RunHarmony}
\usage{
RunHarmony(
  object = NULL,
  batch = "batch",
  dims = 10,
  runningTime = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A seurat object to correct batch effects.}

\item{batch}{Batch labels.}

\item{dims}{Dimensions to use in the correction.}

\item{runningTime}{Return the running time.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
Seurat object with the corrected data in the 'harmony' reduction.
}
\description{
RunHarmony
}
