% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{runQA}
\alias{runQA}
\title{Performs quality checks}
\usage{
runQA(dataFile, preFilter = TRUE, mc.cores = NULL)
}
\arguments{
\item{dataFile}{An R data frame with the data to be processed. The R object is a standard format, and must contain the following headings: File, PE, Sample, Replicate, FilteredFile. More information about the file is available at \code{datafileTemplate}.}

\item{preFilter}{A logical - if true (default), the function will select and analyse files which have not yet been processed for quality. If false, the function will select and analyse those files which have been processed for quality, i.e. the "filtered file" in the data file.}

\item{mc.cores}{A number specifying the number of cores to be used during parallel processing. Default is null - this will result in the number of cores in the default back-end being used. See \code{\url{BioCParallel::registered}}.}
}
\value{
FastqQA object. Outputs quality results in the form of raw data (an R FastqQA object) and HTML format (saved to "QA" directory).
}
\description{
The function checks each fastq file specified in the "data file" for quality, and writes findings to a report.
}
\details{
The function should be run in the working directory, where all fastq files are found.

\code{runQA} iterates over each file specified in the "datafile". It runs a quality assessment from the \code{ShortRead} package. The \code{ShortRead} package (\url{https://bioconductor.org/packages/release/bioc/html/ShortRead.html}) contains more information about this step. The quality assessment may be performed before and after the filtering step, by setting the "pre-filter" parameter to true or to false, respectively. All quality assessment data is output to the "QA" directory. 
Quality reports are output to the working directory (under the QA directory). R objects of the raw data used to generate the reports are also saved to this directory.
}

