% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{runQAandFilter}
\alias{runQAandFilter}
\title{Performs quality checks, then filters reads for quality}
\usage{
runQAandFilter(dataFile, pairedEnd = FALSE, minlength = 30, Phred = 25)
}
\arguments{
\item{dataFile}{An R data frame with the data to be processed. The R object must contain the following headings: File, PE, Sample, Replicate.}

\item{pairedEnd}{A logical. If false (default), a single-end protocol will be run. If true, a paired-end protocol will be run}

\item{Phred}{An integer which specifies Phred (ascii) quality score. Any two consecutive nucleotides with a quality score lower than this threshold will be discarded. Default score is 30}

\item{minLength}{An integer which specifies the minimum length for a read. Reads shorter than this length will be discarded. Default is 30 nucleotides}
}
\value{
An object of class QualityFilterResults. Contains quality checks run before and after the filter, as well as summary statistics of the filtering. Also outputs directories with quality results, and filtered fastq.gz reads
}
\description{
DESCRIPTION GOES HERE
}
\details{
ADD MORE DETAILS
}
\seealso{
\url{https://en.wikipedia.org/wiki/Phred_quality_score} for more about quality scores.
}

