% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCorrection.R
\name{RunScanorama}
\alias{RunScanorama}
\title{RunScanorama}
\usage{
RunScanorama(
  object = NULL,
  batch = "batch",
  runningTime = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A seurat object to correct batch effects.}

\item{batch}{Batch labels.}

\item{runningTime}{Return the running time.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
Seurat object with the corrected data in the 'integrated' assay.
}
\description{
RunScanorama
}
