% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Various.R
\name{SampleData}
\alias{SampleData}
\title{Title}
\usage{
SampleData(object = NULL, frac = NULL, seed = 777, ...)
}
\arguments{
\item{object}{Object to sample}

\item{frac}{Fraction of samples. Must be a value between 0 and 1.}

\item{seed}{Seed to use on sampling}

\item{...}{Arguments passed to other methods.}
}
\value{
Sampled data
}
\description{
Title
}
