% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R, R/qc2.R
\name{SEFilterAndTrim}
\alias{SEFilterAndTrim}
\title{Filters single-end reads for quality}
\usage{
SEFilterAndTrim(file1, minlength = 30, Phred = 25)

SEFilterAndTrim(file1, minlength = 30, Phred = 25)
}
\description{
Filters single-end reads for quality

Filters single-end reads for quality
}

