% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Various.R
\name{SeuratPreprocessing}
\alias{SeuratPreprocessing}
\title{SeuratPreprocessing}
\usage{
SeuratPreprocessing(object = NULL, verbose = FALSE, ...)
}
\arguments{
\item{object}{Seurat object.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
Normalized Seurat object and its variable features.
}
\description{
SeuratPreprocessing
}
