% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice}
\alias{splice}
\title{alternative splicing analysis}
\usage{
splice(pairs, group, colour, bamFiles)
}
\arguments{
\item{pairs}{Vector of length two, either numeric or character, providing the pair of groups to be compared}

\item{group}{Factorial vector with tags for each sample}

\item{colour}{Vector containing color for each condition}

\item{bamFiles}{The path to bam files}
}
\description{
Its an Integrative function for the analysis of alternative splicing. it provides with Alternative splicing discovery using junctions and DE and DU estimation.
}
\examples{

\dontrun{
bamfiles <- c("extdata/A83.bam", "extdata/A84.bam", "extdata/A85.bam", "extdata/A86.bam", "extdata/A94.bam", "extdata/A95.bam")
bamfiles <- c("extdata/SRR5253683.bam", "extdata/SRR5253684.bam", "extdata/SRR5253685.bam",
 "extdata/SRR5253686.bam", "extdata/SRR5253687.bam", "extdata/SRR5253688.bam","extdata/SRR5253689.bam",
 "extdata/SRR5253690.bam","extdata/SRR5253691.bam","extdata/SRR5253692.bam","extdata/SRR5253694.bam",
"extdata/SRR5253695.bam","extdata/SRR5253696.bam","extdata/SRR5253697.bam")
col <- c("red","yellow","green")
col <- c("red","yellow","green","blue", "black", "pink", "violet")
grp <- c(rep("U1", 2),rep("M3", 2),rep("N1", 2))
grp <- c(rep("N1", 2),rep("N2", 2),rep("M1", 2),rep("M2", 2),rep("M3", 2),rep("M4", 2),rep("U1", 2))
splice(pairs = c("U1", "M3"), group = grp, colour = col, bamFiles=bamfiles)
}


\dontrun{
bamfiles <- c("extdata/A83.bam", "extdata/A84.bam", "extdata/A85.bam", "extdata/A86.bam", "extdata/A94.bam", "extdata/A95.bam")
bamfiles <- c("extdata/SRR5253683.bam", "extdata/SRR5253684.bam", "extdata/SRR5253685.bam",
 "extdata/SRR5253686.bam", "extdata/SRR5253687.bam", "extdata/SRR5253688.bam","extdata/SRR5253689.bam",
 "extdata/SRR5253690.bam","extdata/SRR5253691.bam","extdata/SRR5253692.bam","extdata/SRR5253694.bam",
"extdata/SRR5253695.bam","extdata/SRR5253696.bam","extdata/SRR5253697.bam")
col <- c("red","yellow","green")
col <- c("red","yellow","green","blue", "black", "pink", "violet")
grp <- c(rep("U1", 2),rep("M3", 2),rep("N1", 2))
grp <- c(rep("N1", 2),rep("N2", 2),rep("M1", 2),rep("M2", 2),rep("M3", 2),rep("M4", 2),rep("U1", 2))
splice(pairs = c("U1", "M3"), group = grp, colour = col, bamFiles=bamfiles)
}
}
\seealso{
\code{\link{ASpli}}
}
