% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgseaPlot.R
\name{ssgseaPlot}
\alias{ssgseaPlot}
\title{Function "ssgseaPlot"}
\usage{
ssgseaPlot(ssgsea, condition, n1, n2, data_info, p, attr1, attr2,
  conditionI)
}
\arguments{
\item{ssgsea}{an ExpressionSet object containing the result of ssGSEA}

\item{condition}{a string character representing the group condition}

\item{n1}{a numeric variable indicating the number of samples for one of the group condition}

\item{n2}{a numeric variable indicating the number of samples for the other group condition}

\item{data_info}{data frame containing clinical information}

\item{p}{a numeric variable indicating the p value cutoff of wilcoxon test}

\item{attr1}{a string character indicating the column names in clinical file for SampleID}

\item{attr2}{a string character indicating the column containing group information in clinical data}

\item{conditionI}{a character vector of alternative group condition}
}
\value{
a list of plots and p values of wilcoxon test
}
\description{
This function performs single sample GSEA
}
