% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgseaResult.R
\name{ssgseaResult}
\alias{ssgseaResult}
\title{Function "ssgseaResult"}
\usage{
ssgseaResult(ssgseaRes, ssgsea, gs, group, condition, Date)
}
\arguments{
\item{ssgseaRes}{an ExpressionSet object of ssGSEA result}

\item{ssgsea}{result containing all ssGSEA plots object}

\item{gs}{a string character indicating the name of gene set}

\item{group}{a string character indicating the group condition}

\item{condition}{a character vector indicating the group condition}

\item{Date}{a Date object obtained from Sys.Date}
}
\description{
This function stores the ssGSEA results and plots
}
