% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcStats.R
\name{summariseFilteringStats}
\alias{summariseFilteringStats}
\title{Collects summary statistics on filtering}
\usage{
summariseFilteringStats(x)
}
\arguments{
\item{x}{An object of class QualityFilterResults (e.g. returned from \code{runQAandFilter}).}
}
\value{
A dataframe. The dataframe contains the number of reads which have been trimmed (on leading or trailing tails) for poor quality or unknown bases, the number of reads which have been removed entirely due to poort quality, possession of unknown bases, or length too short. Also produces graphs of absolute number of reads output, and relative numbers of reads which have been trimmed or removed (in PDF format).
}
\description{
DESCRIPTION GOES HERE
}
\seealso{
\url{runQAandFilter}
}

