% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{availableReferences}
\alias{availableReferences}
\title{A number of transcriptome references have already been built, for use with \code{kallistoQuant}}
\usage{
availableReferences()
}
\arguments{
\item{None}{No parameters}
}
\value{
A list of available transcriptome indices (.idx files)
}
\description{
A number of transcriptome index files (.idx files) have been pre-built and provided with the package. These files may be used as input to the \code{kallsitoQuant} pseudoalignment function (or the user may provide his/her own transcriptome file). The \code{availableReferences} function provides a list of these ready-available transcriptomes.
}
\examples{
availableReferences()
 
}

