% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblTOentrez.R
\name{ensemblTOentrez}
\alias{ensemblTOentrez}
\title{Function "ensemblTOentrez"}
\usage{
ensemblTOentrez(geneExp, geneSymbol)
}
\arguments{
\item{geneExp}{a data object with ensembl IDs as rownames}

\item{geneSymbol}{a data frame with all gene annotation information}
}
\value{
a list consisting of the corresponding entrezgene IDs and a new data object with entrezgene ID as rownames
}
\description{
This function can be called without DESeq, input a gene expression matrix based on ensembl IDs
and output the corresponding matrix based on entrez IDs
}
