% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{kallistoIndex}
\alias{kallistoIndex}
\title{Builds a transcriptome index from a FASTA formatted file of target sequences}
\usage{
kallistoIndex(refTranscriptome, indexName = "./transcripts.idx")
}
\arguments{
\item{refTranscriptome}{A character string of the transcriptome file (in
fasta format) to be indexed.}

\item{indexName}{A character string of the output index file to be created.
Default is "./transcripts.idx").}
}
\value{
An index file which can be used to quantify expression data using
  \code{\link{kallistoQuant}} pseudoalignment.
}
\description{
DESCRIPTION GOES HERES. Kallisto must be installed locally. Kallisto can be downloaded from \url{https://pachterlab.github.io/kallisto/download.html}.
}
\examples{
kallistoIndex("./transcripts.fasta")
kallistoIndex("./transcripts.fa.gz", "./index.idx")
 
}
\seealso{
\url{http://pachterlab.github.io/kallisto/} for more about Kallisto.
}

