% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{SEFilterAndTrim}
\alias{SEFilterAndTrim}
\title{Filters single-end reads for quality}
\usage{
SEFilterAndTrim(file, minlength = 30, Phred = 25)
}
\arguments{
\item{file}{A character string that is the name of the file to be processed. The input file must be a zipped fastq file (fastq.gz), but this extension should not be included in the character string.}

\item{Phred}{An integer which specifies Phred (ascii) quality score. Any two consecutive nucleotides with a quality score lower than this threshold will be discarded. Default score is 30.}

\item{minLength}{An integer which specifies the minimum length for a read. Reads shorter than this length will be discarded. Default is 30 nucleotides.}
}
\value{
A data frame containing statistics on the QC - how many reads were processed, how many reads were removed, the length of the reads. This information is appended to a "filteringStats" document in the working directory. In addition, the function generates a filtered fastq.gz file ("file-filt.fastq.gz").
}
\description{
DESCRIPTION GOES HERE
}
\examples{
SEFilterAndTrim(file = "reads")
SEFilterAndTrim(file = "reads", minlength = 35, Phred = 30)
SEFilterAndTrim(file = "reads", Phred = 30)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Phred_quality_score} for more about quality scores. \code{\link{PEFilterAndTrim}} processes paired-end data in a similar manner.
}

